# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FermeManager Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sante/vaccins:
    get:
      operationId: getVaccins
      tags: [sante]
      summary: List all vaccinations
      responses:
        "200":
          description: Vaccinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vaccin"
    post:
      operationId: createVaccin
      tags: [sante]
      summary: Record a vaccination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVaccinBody"
      responses:
        "201":
          description: Vaccination recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vaccin"
  /sante/vaccins/{id}:
    delete:
      operationId: deleteVaccin
      tags: [sante]
      summary: Delete a vaccination record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /sante/traitements:
    get:
      operationId: getTraitements
      tags: [sante]
      summary: List all treatments
      responses:
        "200":
          description: Treatments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Traitement"
    post:
      operationId: createTraitement
      tags: [sante]
      summary: Record a treatment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTraitementBody"
      responses:
        "201":
          description: Treatment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Traitement"
  /sante/quarantaine:
    get:
      operationId: getQuarantaine
      tags: [sante]
      summary: List animals in quarantine
      responses:
        "200":
          description: Quarantine list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quarantaine"
    post:
      operationId: createQuarantaine
      tags: [sante]
      summary: Place an animal in quarantine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuarantaineBody"
      responses:
        "201":
          description: Quarantine recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quarantaine"
  /sante/mortalite:
    get:
      operationId: getMortalite
      tags: [sante]
      summary: List mortality records
      responses:
        "200":
          description: Mortality records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mort"
    post:
      operationId: createMort
      tags: [sante]
      summary: Record a death
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMortBody"
      responses:
        "201":
          description: Death recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mort"
...
  /sante/traitements/{id}:
    delete:
      operationId: deleteTraitement
      tags: [sante]
      summary: Delete a treatment record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /sante/quarantaine/{id}:
    delete:
      operationId: deleteQuarantaine
      tags: [sante]
      summary: Delete a quarantine record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /sante/mortalite/{id}:
    delete:
      operationId: deleteMort
      tags: [sante]
      summary: Delete a mortality record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /utilisateurs:
    get:
      operationId: getUtilisateurs
      tags: [utilisateurs]
      summary: List all users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Utilisateur"
    post:
      operationId: createUtilisateur
      tags: [utilisateurs]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUtilisateurBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Utilisateur"
  /utilisateurs/{id}:
    put:
      operationId: updateUtilisateur
      tags: [utilisateurs]
      summary: Update a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUtilisateurBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Utilisateur"
    delete:
      operationId: deleteUtilisateur
      tags: [utilisateurs]
      summary: Delete a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── SYSTEM SETTINGS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Vaccin:
      type: object
      properties:
        id:
          type: integer
        tag:
          type: string
        vaccin:
          type: string
        date:
          type: string
        dose:
          type: string
          nullable: true
        rappel:
          type: string
          nullable: true
        administrePar:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, tag, vaccin, date, createdAt]
    CreateVaccinBody:
      type: object
      properties:
        tag:
          type: string
        vaccin:
          type: string
        date:
          type: string
        dose:
          type: string
          nullable: true
        rappel:
          type: string
          nullable: true
        administrePar:
          type: string
          nullable: true
      required: [tag, vaccin, date]
    Traitement:
      type: object
      properties:
        id:
          type: integer
        tag:
          type: string
        typeTraitement:
          type: string
        produit:
          type: string
        dose:
          type: string
          nullable: true
        dateDebut:
          type: string
        dateFin:
          type: string
          nullable: true
        statut:
          type: string
        createdAt:
          type: string
      required: [id, tag, typeTraitement, produit, dateDebut, statut, createdAt]
    CreateTraitementBody:
      type: object
      properties:
        tag:
          type: string
        typeTraitement:
          type: string
        produit:
          type: string
        dose:
          type: string
          nullable: true
        dateDebut:
          type: string
        dateFin:
          type: string
          nullable: true
        statut:
          type: string
      required: [tag, typeTraitement, produit, dateDebut, statut]
    Quarantaine:
      type: object
      properties:
        id:
          type: integer
        tag:
          type: string
        motif:
          type: string
        dateDebut:
          type: string
        dureeJours:
          type: integer
        statut:
          type: string
        createdAt:
          type: string
      required: [id, tag, motif, dateDebut, dureeJours, statut, createdAt]
    CreateQuarantaineBody:
      type: object
      properties:
        tag:
          type: string
        motif:
          type: string
        dateDebut:
          type: string
        dureeJours:
          type: integer
      required: [tag, motif, dateDebut, dureeJours]
    Mort:
      type: object
      properties:
        id:
          type: integer
        tag:
          type: string
        date:
          type: string
        cause:
          type: string
        confirme_par:
          type: string
          nullable: true
        observations:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, tag, date, cause, createdAt]
    CreateMortBody:
      type: object
      properties:
        tag:
          type: string
        date:
          type: string
        cause:
          type: string
        confirme_par:
          type: string
          nullable: true
        observations:
          type: string
          nullable: true
      required: [tag, date, cause]
...
    Utilisateur:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
        role:
          type: string
        modules:
          type: array
          items:
            type: string
        actif:
          type: boolean
        createdAt:
          type: string
      required: [id, username, nom, prenom, email, role, modules, actif, createdAt]
    CreateUtilisateurBody:
      type: object
      properties:
        username:
          type: string
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
        role:
          type: string
        modules:
          type: array
          items:
            type: string
        actif:
          type: boolean
        password:
          type: string
      required: [username, nom, prenom, email, role, password]
    UpdateUtilisateurBody:
      type: object
      properties:
        username:
          type: string
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
        role:
          type: string
        modules:
          type: array
          items:
            type: string
        actif:
          type: boolean
        password:
          type: string
          nullable: true
...